style: login-signup basic template
