+++ apps/commudle-admin/src/app/components/login/login.component.html
@@ -1,53 +1,7 @@
 <div class="base-layout thin">
   <div class="container md:!com-w-[530px]">
     <div class="one-column-layout">
-      <nb-card class="login-container">
-        <nb-card-body class="com-self-center com-space-y-8">
-          <h2>Login/Signup with Commudle</h2>
-
-          <div
-            *ngIf="!libAuthWatchService.getAuthCookie()"
-            class="com-flex com-justify-center"
-            google-signin-button
-            type="standard"
-          ></div>
-
-          <form [formGroup]="loginForm" class="login-form">
-            <input formControlName="email" fullWidth nbInput placeholder="Ex. more@commudle.com" type="email" />
-            <input
-              *ngIf="isEmailSent"
-              formControlName="code"
-              fullWidth
-              nbInput
-              placeholder="Enter OTP here"
-              type="password"
-              status="info"
-            />
-            <button
-              (click)="sendVerificationEmail()"
-              *ngIf="!isEmailSent"
-              [disabled]="loginForm.get('email').invalid"
-              [nbSpinner]="isLoading"
-              fullWidth
-              nbButton
-              status="primary"
-            >
-              Send Verification Code
-            </button>
-            <button
-              (click)="loginUser()"
-              *ngIf="isEmailSent"
-              [disabled]="loginForm.invalid"
-              [nbSpinner]="isLoading"
-              fullWidth
-              nbButton
-              status="primary"
-            >
-              Login
-            </button>
-          </form>
-        </nb-card-body>
-      </nb-card>
+      <commudle-login-signup></commudle-login-signup>
     </div>
   </div>
 </div>
